backEnd/requirements.txt
astroid==3.2.2
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dill==0.3.8
filelock==3.15.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.7
importlib-metadata==4.11.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.8
mysql-connector==2.2.9
mysql-connector-python==8.0.29
pip==22.0.2
platformdirs==4.2.2
protobuf==4.21.1
PyJWT==2.8.0
pylint==3.2.3
pylint-venv==3.0.3
pynvim==0.5.0
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.32.3
requests-file==2.1.0
setuptools==59.6.0
tldextract==5.1.2
tomli==2.0.1
tomlkit==0.12.5
typing-extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.7.0
zope.event==5.0
zope.interface==6.4.post2
